+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 All changes to the package starting with v2.8.1 will be logged here.
 
-## v2.24.0 [2024-01-XX]
+## v2.24.0rc1 [2024-01-XX]
 #### What's New
-* None
+* `britive.settings.banner.*` - administer the banner/system announcement
+* `britive.my_access.banner` - view the banner/system announcement (all end users can view the banner)
 
 #### Enhancements
 * Implement logic to catch and present user-friendly error if a tenant is under maintenance